db-schema: convert blobs to typescript
