+++ src/smc-util/db-schema/account-profiles.ts
@@ -1,6 +1,5 @@
 import { create } from "./types";
 
-
 export const account_profiles = create({
   rules: {
     desc:
